// services/admin_portal/src/app/models/branch.model.ts
export interface Branch {
  branchId?: number,
  branchCode: string;
  branchName: string;
  branchManager: number;
  phoneNumber: string;
  email: string;
  address1: string;
  address2: string;
  city: string;
  state: string;
  postalCode: string;
  country: string;
  lat: number;
  lng: number;
}